adding print to consignment, pos, and custom order

+++ models/domain/sales_return_detail.go
@@ -9,7 +9,7 @@ type SalesReturnDetail struct {
 	ProductId int `json:"product_id"`
 	Qty int `json:"qty"`
 	Total float64 `json:"total"`
-	UnitPrice int `json:"unit_price"`
+	UnitPrice float64 `json:"unit_price"`
 	CreatedAt time.Time`json:"created_at,omitempty" example:"2020-03-16T13:55:09.598136+07:00"`
 	UpdatedAt time.Time `json:"updated_at,omitempty" example:"2020-03-16T13:55:09.598136+07:00"`
 }
